+++ lib/paco/rspec.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+# Utils for testing Paco
+require "paco/rspec/parse_matcher"